Move execution code into separate file
